+++ src/components/Accordion/Accordion.styles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const AccordionLabel = styled.div`
+  margin-top: 1rem;
+  padding: 0 0.6rem 0.6rem;
+  border-bottom: 1px solid #ddd;
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: stretch;
+  width: 100%;
+  font-size: 1rem;
+  font-weight: bold;
+  text-align: left;
+  color: #007bc7;
+`
