+++ src/components/Accordion/presentational/List.tsx
@@ -2,45 +2,45 @@ import React, { useState } from 'react'
 import styled from 'styled-components'
 
 const ClassWrapperStyle = styled.div`
-    border-bottom: 1px solid #ddd;
+  border-bottom: 1px solid #ddd;
 `
 
 const ClassStyle = styled.div`
-    display: flex;
-    justify-content: space-between;
-    align-items: stretch;
+  display: flex;
+  justify-content: space-between;
+  align-items: stretch;
 
-    &:hover {
-        background-color: #f7f7f7;
-    }
+  &:hover {
+    background-color: #f7f7f7;
+  }
 `
 
 const ClassNameStyle = styled.div`
-    padding: 0.65rem;
-    width: 100%;
-    color: #333;
-    text-decoration: none;
-    text-align: left;
-    cursor: pointer;
+  padding: 0.65rem;
+  width: 100%;
+  color: #333;
+  text-decoration: none;
+  text-align: left;
+  cursor: pointer;
 `
 
 const ClassCountStyle = styled.small`
-    color: #999;
+  color: #999;
 `
 
 const ClassListStyle = styled.div`
-    transition: height 0.15s ease-in-out;
-    overflow: hidden;
+  transition: height 0.15s ease-in-out;
+  overflow: hidden;
 `
 
-const AccordionWrapper = ({
+const List = ({
   title,
   itemLength,
   children,
 }: {
-    title: string
-    itemLength: number
-    children: React.ReactChild | Array<React.ReactChild>
+  title: string;
+  itemLength: number;
+  children: React.ReactChild | Array<React.ReactChild>;
 }) => {
   const [isAccordionOpened, setIsAccordionOpened] = useState(false)
 
@@ -61,4 +61,4 @@ const AccordionWrapper = ({
   )
 }
 
-export default AccordionWrapper
+export default List
