+++ src/components/Button/Button.stories.tsx
@@ -1,6 +1,6 @@
 import { action } from '@storybook/addon-actions'
 
-import { Button } from '../components/button'
+import { Button } from './Button'
 
 export const ButtonStory = () => (
   <Button onClick={action('clicked')}>Button</Button>
