+++ src/components/Button/Button.styles.tsx
@@ -1,0 +1,16 @@
+import styled from 'styled-components'
+
+export const ButtonStyle = styled.button`
+  display: inline-flex;
+  align-items: center;
+  padding: 0.8rem;
+  text-decoration: none;
+  text-align: center;
+  border: 1px solid #007bc7;
+  border-radius: 4px;
+  cursor: pointer;
+  line-height: 1;
+  white-space: nowrap;
+  background-color: #fff;
+  color: #000;
+`
