+++ src/components/Input/Input.spec.tsx
@@ -3,34 +3,17 @@ import Enzyme, { shallow } from 'enzyme'
 
 Enzyme.configure({ adapter: new Adapter() })
 
-import { Select as NekoSelect } from '../src/components/select'
-
-const options = [
-  {
-    value: 1,
-    text: 'Angular',
-  },
-  {
-    value: 2,
-    text: 'React',
-  },
-  {
-    value: 3,
-    text: 'Vue',
-  },
-]
+import { Input as NekoInput } from './Input'
 
 const setValue = () => {
   console.log('updated')
 }
 
-describe('Select', () => {
+describe('Input', () => {
   let wrapper
 
   beforeEach(() => {
-    wrapper = shallow(
-      <NekoSelect options={options} value={1} onChange={setValue} />
-    )
+    wrapper = shallow(<NekoInput value={'Input'} onChange={setValue} />)
   })
 
   it('should render without errors', () => {
