+++ src/components/Label/Label.styles.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components'
+
+export const LabelStyle = styled.div`
+  display: inline-block;
+  padding: 0.5rem 0.75rem;
+  margin: 0.25rem;
+  background-color: #68788d;
+  color: #fff;
+  border-radius: 9999px;
+  border-width: 1px;
+  border-color: #2e3439;
+  a {
+    color: #68788d;
+  }
+`
