+++ src/index.ts
@@ -1,8 +1,8 @@
 // Atomic Components
-export { Button as NekoButton } from './components/button'
-export { Input as NekoInput } from './components/input'
-export { Label as NekoLabel } from './components/label'
-export { Select as NekoSelect } from './components/select'
+export { Button as NekoButton } from './components/Button/Button'
+export { Input as NekoInput } from './components/Input/Input'
+export { Label as NekoLabel } from './components/Label/Label'
+export { Select as NekoSelect } from './components/Select/Select'
 
 // Modules
-export { Accordion as NekoAccordion } from './components/accordion'
+export { Accordion as NekoAccordion } from './components/Accordion/Accordion'
